refactor: split the capture machinery into capture_store and hf_dataset
The write-capturing object store is generic — it decorates any inner store and
records writes instead of forwarding them — so it now stands on its own in
capture_store.rs (Lance-free, reusable, with its unit tests). hf_dataset.rs keeps
the HF/Lance layer: the WrappingObjectStore seam plus the append/reindex
operations that run a native Lance op through the capture store and return the
files it wrote. The public surface is hf_dataset::append / hf_dataset::reindex;
"capture" stays an implementation detail on the private helpers and types.
