+++ src/lib.rs
@@ -5,9 +5,10 @@
 //! The binary ([`main`]) is a thin CLI over these modules; integration tests drive
 //! them directly.
 
-pub mod capture;
+pub mod capture_store;
 pub mod chunk;
 pub mod db;
+pub mod hf_dataset;
 pub mod hub;
 pub mod index;
 pub mod mcp;
